Bump using_directives version. Move using directives logic to separate object
