+++ scaladoc/src/dotty/tools/scaladoc/snippets/SnippetChecker.scala
@@ -11,23 +11,12 @@ import dotty.tools.dotc.fromtasty.TastyFileUtil
 import dotty.tools.dotc.config.Settings._
 import dotty.tools.dotc.config.ScalaSettings
 
-import com.virtuslab.using_directives._
-import com.virtuslab.using_directives.custom.model._
-import com.virtuslab.using_directives.custom.utils.{ Position => UDPosition }
-import com.virtuslab.using_directives.reporter._
 import scala.util.chaining._
 
 class SnippetChecker(val args: Scaladoc.Args)(using cctx: CompilerContext):
   import SnippetChecker._
 
-// (val classpath: String, val bootclasspath: String, val tastyFiles: Seq[File], isScalajs: Boolean, useJavaCp: Boolean):
   private val sep = System.getProperty("path.separator")
-
-  private val usingDirectivesProcessor = {
-    val processor = UsingDirectivesProcessor()
-    // TODO: Implement custom reporter
-    processor
-  }
 
   private val fullClasspath = List(
     args.tastyFiles
@@ -60,22 +49,9 @@ class SnippetChecker(val args: Scaladoc.Args)(using cctx: CompilerContext):
     val outerLineOffset = lineOffset + data.fold(0)(_.position.line) + constantLineOffset
     val outerColumnOffset = data.fold(0)(_.position.column) + constantColumnOffset
 
-    val udReporter = CustomUsingDirectivesReporter(sourceFile, outerLineOffset, outerColumnOffset)
-    usingDirectivesProcessor.getContext.setReporter(udReporter)
-
-    val importsWithUsingDirectives: Seq[(Import, UsingDirectives)] = snippetImports.map(i => i -> extractUsingDirectives(i.snippet))
-    val truncatedImports = importsWithUsingDirectives.map {
-      case (i: Import, ud: UsingDirectives) => i.copy(snippet = i.snippet.substring(ud.getCodeOffset()))
-    }
-    val importUds = importsWithUsingDirectives.map(_(1))
-    val snippetUd = extractUsingDirectives(snippet)
-    val truncatedSnippet = snippet.substring(snippetUd.getCodeOffset())
-    val mergedUsingDirectives = mergeUsingDirectives(importUds :+ snippetUd)
-
-    val (additionalCompilerSettings, parsingArgsMessages) = UsingDirectivesSettingsExtractor.process(mergedUsingDirectives)
-
-    // We can provide more accurate position of error by looking for error in AST
-    val additionalMessage = Option.when(!parsingArgsMessages.isEmpty)(createParsingReportMessage(parsingArgsMessages, outerLineOffset, sourceFile))
+    val (truncatedSnippet, truncatedImports, additionalCompilerSettings, additionalMessages) = UsingDirectivesExtractor.processSnippet(
+      snippet, snippetImports, outerLineOffset, outerColumnOffset, sourceFile
+    )
 
     val mergedSnippet = mergeSnippets(truncatedSnippet, truncatedImports)
     if arg.flag != SCFlags.NoCompile then
@@ -89,7 +65,7 @@ class SnippetChecker(val args: Scaladoc.Args)(using cctx: CompilerContext):
         additionalCompilerSettings
       )
 
-      val res = compiler.compile(wrapped, arg, sourceFile).pipe(result => result.copy(messages = result.messages ++ additionalMessage ++ udReporter.getMessages))
+      val res = compiler.compile(wrapped, arg, sourceFile).pipe(result => result.copy(messages = result.messages ++ additionalMessages))
       Some(res)
     else None
   }
@@ -103,58 +79,6 @@ class SnippetChecker(val args: Scaladoc.Args)(using cctx: CompilerContext):
     }
     wrappedImports :+ snippet
   }.mkString("\n")
-
-  private def mergeUsingDirectives(uds: Seq[UsingDirectives]): Map[Path, Value[_]] = {
-    import collection.JavaConverters._
-    val maps = uds.map(_.getFlattenedMap.asScala.toMap)
-    maps.reduceLeft {
-      case (map1, map2) => map1 ++ map2
-    }
-  }
-
-  private def extractUsingDirectives(snippet: String): UsingDirectives =
-    usingDirectivesProcessor.extract(snippet.toCharArray)
-
-  private def createParsingReportMessage(msgs: Seq[String], lineOffset: Int, sourceFile: SourceFile): SnippetCompilerMessage = {
-    import dotty.tools.dotc.util.Spans._
-    val offsetFromLine = sourceFile.lineToOffset(lineOffset - 1)
-    val span = Span(offsetFromLine, offsetFromLine)
-    val pos = Some(Position(dotty.tools.dotc.util.SourcePosition(sourceFile, span), 0))
-    SnippetCompilerMessage(pos, msgs.mkString("\n"), MessageLevel.Warning)
-  }
-
-  private class CustomUsingDirectivesReporter(sourceFile: SourceFile, outerLineOffset: Int, outerColumnOffset: Int) extends Reporter:
-    import dotty.tools.dotc.util.Spans._
-
-    private val builder = Seq.newBuilder[SnippetCompilerMessage]
-
-    private def buildMessage(level: MessageLevel, msg: String, position: Option[UDPosition] = None) = {
-      val pos = position.map { p =>
-        val offsetFromLine = sourceFile.lineToOffset(p.getLine() + outerLineOffset - 1)
-        val offsetFromColumn = p.getColumn()
-        val span = Span(
-          offsetFromLine + offsetFromColumn + outerColumnOffset,
-          offsetFromLine + offsetFromColumn + outerColumnOffset
-        )
-        Position(dotty.tools.dotc.util.SourcePosition(sourceFile, span), 0)
-      }
-      SnippetCompilerMessage(pos, msg, level)
-    }
-
-    def getMessages: Seq[SnippetCompilerMessage] =
-      builder.result()
-
-    override def error(msg: String) = builder += buildMessage(MessageLevel.Warning, msg)
-
-    override def warning(msg: String) = builder += buildMessage(MessageLevel.Warning, msg)
-
-    override def error(pos: UDPosition, msg: String) = builder += buildMessage(MessageLevel.Warning, msg, Some(pos))
-
-    override def warning(pos: UDPosition, msg: String) = builder += buildMessage(MessageLevel.Warning, msg, Some(pos))
-
-    override def hasErrors(): Boolean = builder.knownSize != 0
-
-    override def reset(): Unit = { } // We recreate reporter each time so its not needed to reset anything
 
 
 object SnippetChecker:
